refactor: #12 UserServiceTest 수정
